API

We can summarize the main functions of this package as follows.

Types

MolecularConformation.NMRInfo — Type

NMRinfo :: Type

This type contains the information for NMRtype. In this point of the project it contains two fields: dist, atom1 and atom2. Essentially, dist is the distance between atom1 and atom2 which are strings related to the name of atoms.

MolecularConformation.NMRType — Type

NMRType :: Type

This type is used as input data. Essentially, this type carry itself fields for internal solvers. Although it is possible to input data directly into the type, it is highly recommended that this type be constructed from the function preprocessing, like in the following example

Example

nmr_data_example = preprocessing("1A57.nmr")

As return an element of NMRtype is created. We are assuming that a 1A57.nmr file is provided.

MolecularConformation.AtomType — Type

AtomType :: Type

It is the most primitive type and it is used to store the spatial positions of atoms. It is a mutable type. This type has three field: .x, .y and .z

MolecularConformation.MoleculeType — Type

MoleculeType :: Type

It is used to store a (just one) solution find by conformation process. It is a mutable struct and has two fields: .atoms (used to store a vector of AtomType) and .lde (used to store the lde of an molecule)

MolecularConformation.ConformationSetup — Type

ConformationSetup :: Type

This type is used to define the main options of the function used to conformation. Essentially, this type has three fields: precision, solver, allsolutions.

precision :: is associated to current internal solver which depends on some level do a pruning test.

solver :: is the name of function which define the solver.

allsolution :: is a bool variable used to define the behavior of the method. Is set as true, the internal solver try to find all possible solutions.

Example

options = ConformationSetup(0.00001,classicBP,true)

As return an element options was created an it will used in conformation function. It is a mandatory definition before run conformation function.

MolecularConformation.ConformationOutput — Type

ConformationOutput :: Type

It is a mutable type used to store the output provided by conformation function. With this type is possible to handle with some importants elements given by the following fields:.number,.molecules,.elapsedtime,.bytes and .gctime

Main functions

MolecularConformation.preprocessing — Function

preprocessing :: Function

This function is able to read a .nmr file and generate a NMRType, which is an internal type with fields that can be used by internals solvers. Given a list in .nmr format, let us say the 1A57.nmr file, the following example show how simple is to use this function.

Example

preprocessing("1A57.nmr")

MolecularConformation.conformation — Method

conformation(NMRdata,cs::ConformationSetup)

This is the main function in order to get the conformation of a molecule. To run this function, we need to setup a NMR file and then define some options using the ConformationSetup type.

Example

julia> options = ConformationSetup(0.001,classicBP,true)

julia> conformation(NMRdata,options)

as return a ConformationOutput type is provided.

There are others parameters to setup, for example, need to complete.

MolecularConformation.classicBP — Method

classicBP :: Function

This function is an important solver of this package. This solver implements the algorithm given in

Liberti, L., Lavor, C., & Maculan, N. (2008). A branch‐and‐prune algorithm for the molecular distance geometry problem. International Transactions in Operational Research, 15(1), 1-17.

The main difference of our implementation is that the input data (NMRType) can store informations of preprocessing function and as consequence we can optimize the search tree of this algorithm.

MolecularConformation.LDE — Method

LDE :: Function

This function is an auxiliary function used to compute the LDE of some molecule. This function change the .lde field of the molecule.

MolecularConformation.build_distance_matrix — Method

build_distance_matrix

This function can be useful to visualize a the distance array associated to an MoleculeType solution

Example

v = conformation(data,opt)
build_distance_array(v.molecules[1].atoms)

MolecularConformation.convert_to_dataframe — Method

convert_to_dataframe

A function used to convert an array of AtomType elements in a dataframe

Examples

julia-repl
julia> sol = conformation(data,options) # assuming all solutions founded

julia> convert_to_dataframe(sol.molecules[1].atoms)

returns a dataframe object. 

Secundary functions

This functions are used as auxiliary functions by internal solvers.

MolecularConformation.generate_virtual_path — Method

generate_virtual_path :: Function

This function is an auxiliary function used to define a useful vector called in our context as virtual path. This vector allows to handle with re-order approach.

MolecularConformation.bondangle — Method

bondangle :: Function

This is an auxiliary function used by ClassicBP solver in order to compute the bond angle by cosine rule. As output the cosine and sine are given.

MolecularConformation.torsionangle — Method

torsionangle :: Function

This is an auxiliary function used by classicBPOpt solver in order to compute the torsion angle. As output the cosine and sine of the torsion angle are given.

MolecularConformation.badtorsionangle — Method

badtorsionangle :: Function

This is an auxiliary function used by classicBP solver in order to compute the torsion angle. As output the cosine and sine of the torsion angle are given.

MolecularConformation.torsionmatrix — Method

torsionmatrix :: Function

This is an auxiliary function used by classicBPOpt (optimization version of classicBP) solver to compute the torsion array.

MolecularConformation.pruningtest — Method

pruningtest :: Function

This functions is an auxiliary function used to test if some molecule is feasible or not.

MolecularConformation.outputfilter — Function

outputfilter

This function is especially useful when we need to filter an information about a solution in ConformationOutput. For example, let us suppose that we want to compute the worst LDE measure. So, we need to type:

julia> a = conformation(data,option)

julia> outputfilter(a,"lde")